chore(zbugs): rename `push.ts` to deal with `nextjs` routing
`api/push` is causing nextjs to load `push.js` directly instead of using `index.js`
